refactor(cli): replace build/plan commands with unified run command
Consolidate the separate build and plan commands into a single run
command. This simplifies the CLI interface while maintaining all
functionality. Users now use 'goralph run' instead of 'goralph build'
or 'goralph plan'.

BREAKING CHANGE: The 'build' and 'plan' subcommands have been removed.
Use 'goralph run' instead.
